requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
pymysql==1.1.1
python-dateutil==2.8.2
flask-cors==4.0.0
flask-restx==1.3.0
hl7apy==1.3.4
fhir.resources>=7.0.0
PyJWT==2.8.0